Configured button box for this year. No buttons implemented yet.
